feat: Initial translate API (#15)

+++ runtime/functor-runtime-common/src/frame_time.rs
@@ -3,8 +3,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Serialize, Deserialize)]
 pub struct FrameTime {
     // tts - total time in seconds
-    pub tts: f64,
+    pub tts: f32,
 
     // dts - delta time in seconds for this frame
-    pub dts: f64,
+    pub dts: f32,
 }
